requirements.txt
matplotlib==3.6.2
numpy==1.23.4
pandas==1.4.2
Pillow==9.3.0
seaborn==0.12.1
streamlit==1.14.0
